Hacky fix for replacing template content properly
WordPress seems to offer no proper info about which editor field
is being manipulated when using the `the_editor_content` filter.

A hacky workaround is to see how many times a replacement has been
made in a filter callback so we replace field contents only for the
very first editor field on each page.

Maybe find some other solution for this, not sure if anyone is going
to insert editor fields above the stock field somewhere.